+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## [In Development](https://github.com/ppfeufer/eve-online-fitting-manager/tree/development)
 [Full Changelog](https://github.com/ppfeufer/eve-online-fitting-manager/compare/v0.13.4...HEAD)
 This is a complete rewrite to utilize ESI, so there is no longer a need for a local EDK installation in order to get this plugin to work. Before you update, it is recommended that you backup your fittings you have already.
+### Added
+- ESI Client
+- Database cache table for market, so we don't pollute the wp_options table to much
 
 ### Changed
 - EVE Central market API deactivated, since it's is still dead